refactor: separate customer behaviours from its entity
